Replace @type state by detecting an array when generating the schema
The previous `@type = :array` logic was to detect if an array was set
within a `_scope` block, however we can omit that and detect the array
types after the fact. Hereby checking if our @attributes are an array
or a nested set is (though we only check the top-level for now).

As a sidenote: when Jbuilder#array! is called the internal @attributes
Hash is swapped out to be an Array instead. However, Jbuilder::Schema
converts that to another Hash so we need extra assurances on top of that.

It may be viable to let the Array through and convert them later when
finalizing the schema in `schema!`.