Hacky bug fix for Rails 7.1.0
Fixes #54 (but I'm not sure it fixes it in a good way).

It looks like something about `ActionView` changed with Rails `7.1.0`,
but I haven't been able to spot exactly what from the Changelog.

I'm not confident that this is a fix that we actually want to use. I'm
mainly submitting this PR to document what I've found, and I'm hoping
that someone will have an idea for a better fix.